feat: added support for evaluating expressions
This commit introduces an *expression* language to AnalysisLang.
The expression language enables terms that still need to be evaluated
to obtain its final term value.

The current iteration of the expression language contains the following
syntactic elements:

* Map lookup: enables looking up values from a "TermMap" term.
* Map update enables updating bindings from a "TermMap" term.

To this end, a number of changes are introduced:

* The AST now contains support for this new expression language,
including references from the term language to the expression language.
* The type checker successfully type checks terms in this expression
language. The type checker also transforms functor applications
to map lookups if the type permits so.
* The unification engine evaluates each expression fully if encountered.
This behavior is similar to the execution of embedded Haskell expressions.