feat(unification): evaluate rewrite rule applications
This commit introduces support for evaluating rewrite rules.
To do so, a number of components of the system are affected:

* The AST has been updated to support any type of term in the
body of the a rewrite rule instead of only pure terms.
* A number of stateful fields have been moved from the solver
to the unifier, the solver now also uses the unifier as its
base monad instead of executing the monad each time a unification
has to be performed.
*  The unification engine has been updated to support rewrite applications
in its "evaluateExpr" function, the ground terms required is still there.